Move EstimatedFeeToMaxFee.kt to account package
