+++ lib/src/main/kotlin/com/swmansion/starknet/account/StandardAccount.kt
@@ -1,6 +1,5 @@
 package com.swmansion.starknet.account
 
-import com.swmansion.starknet.crypto.estimatedFeeToMaxFee
 import com.swmansion.starknet.data.types.*
 import com.swmansion.starknet.data.types.transactions.*
 import com.swmansion.starknet.data.types.transactions.DeployAccountTransactionPayload
